+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Meta_data dependancies [#125](https://github.com/rokwire/content-building-block/issues/125) 
+
 ## [1.13.1] - 2024-10-07
 ### Fixed
 - Fixed environment variable [#127](https://github.com/rokwire/content-building-block/issues/127)